Added workflows for both
