+++ job_orchestrator/host.json
@@ -1,0 +1,8 @@
+{
+  "version": "2.0",
+  "extensionBundle": {
+    "id": "Microsoft.Azure.Functions.ExtensionBundle",
+    "version": "[4.*, 5.2.0)"
+  },
+  "functionTimeout": "00:10:00"
+}
