+++ job_orchestrator/pom.xml
@@ -1,0 +1,107 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
+	<modelVersion>4.0.0</modelVersion>
+
+	<parent>
+		<groupId>com.nomad.multi</groupId>
+		<artifactId>data-services</artifactId>
+		<version>1.0</version>
+	</parent>
+
+	<groupId>com.nomad.</groupId>
+	<artifactId>job_orchestrator</artifactId>
+	<version>0.0.1-SNAPSHOT</version>
+	<name>job_orchestrator</name>
+	<description>Deployed to Azure Function to create web scraping jobs</description>
+
+	<properties>
+		<start-class>com.nomad.producer.JobOrchestratorApplication</start-class>
+		<version.spring.cloud>2024.0.0</version.spring.cloud>
+		<!-- Plugins -->
+		<functionResourceGroup>rg-dev-uks-nomad-02</functionResourceGroup>
+		<functionAppName>fa-dev-uks-nomad-02-job-orchestrator</functionAppName>
+	</properties>
+
+	<dependencyManagement>
+		<dependencies>
+			<!-- Used for spring-cloud-function-adapter-azure below -->
+			<dependency>
+				<groupId>org.springframework.cloud</groupId>
+				<artifactId>spring-cloud-dependencies</artifactId>
+				<version>${version.spring.cloud}</version>
+				<type>pom</type>
+				<scope>import</scope>
+			</dependency>
+		</dependencies>
+	</dependencyManagement>
+
+	<dependencies>
+		<dependency>
+			<groupId>com.nomad.library</groupId>
+			<artifactId>common</artifactId>
+			<version>0.0.1-SNAPSHOT</version>
+		</dependency>
+		
+		<!-- Used to create Azure Functions -->
+		<dependency>
+			<groupId>org.springframework.cloud</groupId>
+			<artifactId>spring-cloud-function-adapter-azure</artifactId>
+		</dependency>
+
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-starter-data-neo4j</artifactId>
+		</dependency>
+
+		<!-- Used to read YML file -->
+		<dependency>
+			<groupId>com.fasterxml.jackson.dataformat</groupId>
+			<artifactId>jackson-dataformat-yaml</artifactId>
+		</dependency>
+
+		<!-- Used to access Key Vault to pull Neo4j password -->
+		<dependency>
+			<groupId>com.azure</groupId>
+			<artifactId>azure-security-keyvault-secrets</artifactId>
+		</dependency>
+	</dependencies>
+
+	<build>
+		<plugins>
+			<plugin>
+				<!-- Azure Functions Maven Plugin to package the Azure Function -->
+				<groupId>com.microsoft.azure</groupId>
+				<artifactId>azure-functions-maven-plugin</artifactId>
+				<version>1.37.0</version>
+				<configuration>
+					<resourceGroup>${functionResourceGroup}</resourceGroup>
+					<appName>${functionAppName}</appName>
+					<hostJson>${project.basedir}/host.json</hostJson>
+                    <localSettingsJson>${project.basedir}/local.settings.json</localSettingsJson>
+				</configuration>
+				<executions>
+					<execution>
+						<goals>
+							<goal>package</goal> <!-- This makes it trigger during 'mvn package' -->
+						</goals>
+					</execution>
+            	</executions>
+			</plugin>
+			<plugin>
+				<groupId>org.apache.maven.plugins</groupId>
+				<artifactId>maven-compiler-plugin</artifactId>
+				<configuration>
+					<annotationProcessorPaths>
+						<path>
+							<groupId>org.projectlombok</groupId>
+							<artifactId>lombok</artifactId>
+						</path>
+					</annotationProcessorPaths>
+				</configuration>
+			</plugin>
+
+		</plugins>
+	</build>
+
+</project>
