+++ job_orchestrator/src/main/java/com/nomad/job_orchestrator/JobOrchestratorApplication.java
@@ -1,0 +1,14 @@
+package com.nomad.job_orchestrator;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class JobOrchestratorApplication {
+
+	public static void main(String[] args) {
+
+		SpringApplication.run(JobOrchestratorApplication.class, args);
+	}
+
+}
