+++ job_orchestrator/src/main/java/com/nomad/job_orchestrator/domain/PotentialRoute.java
@@ -1,0 +1,5 @@
+package com.nomad.job_orchestrator.domain;
+
+import com.nomad.library.domain.City;
+
+public record PotentialRoute(City sourceCity, City destinationCity) {}
