+++ job_orchestrator/src/main/java/com/nomad/job_orchestrator/exceptions/InvalidJobConfig.java
@@ -1,0 +1,9 @@
+package com.nomad.job_orchestrator.exceptions;
+
+public class InvalidJobConfig extends Exception {
+    
+  public InvalidJobConfig(String reason) {
+    super("Job config error; " + reason);
+  }
+
+}
