+++ job_orchestrator/src/main/resources/jobs-config.yml
@@ -1,0 +1,6 @@
+jobs:
+  - id: thailand-routes
+    type: ROUTE_DISCOVERY
+    countryName: Thailand
+    timeToFetch: +2d
+    isActive: true
