--- one2goasia/src/main/java/com/nomad/one2goasia/ConsumerApplication.java
@@ -1,13 +1,0 @@
-package com.nomad.one2goasia;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ConsumerApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ConsumerApplication.class, args);
-	}
-
-}
