+++ one2goasia/src/main/java/com/nomad/one2goasia/One2GoAsiaApplication.java
@@ -1,0 +1,13 @@
+package com.nomad.one2goasia;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class One2GoAsiaApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(One2GoAsiaApplication.class, args);
+	}
+
+}
