+++ pom.xml
@@ -22,7 +22,7 @@
 	<modules>
 		<module>common</module>
         <module>one2goasia</module>
-        <module>producer</module>
+        <module>job_orchestrator</module>
     </modules>
 
     <properties>
