--- producer/src/main/java/com/nomad/producer/ProducerApplication.java
@@ -1,14 +1,0 @@
-package com.nomad.producer;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ProducerApplication {
-
-	public static void main(String[] args) {
-
-		SpringApplication.run(ProducerApplication.class, args);
-	}
-
-}
