--- producer/src/main/java/com/nomad/producer/domain/PotentialRoute.java
@@ -1,5 +1,0 @@
-package com.nomad.producer.domain;
-
-import com.nomad.library.domain.City;
-
-public record PotentialRoute(City sourceCity, City destinationCity) {}
