--- producer/src/main/java/com/nomad/producer/exceptions/InvalidJobConfig.java
@@ -1,9 +1,0 @@
-package com.nomad.producer.exceptions;
-
-public class InvalidJobConfig extends Exception {
-    
-  public InvalidJobConfig(String reason) {
-    super("Job config error; " + reason);
-  }
-
-}
