--- producer/src/main/java/com/nomad/producer/exceptions/ServiceBusProcessingException.java
@@ -1,5 +1,0 @@
-package com.nomad.producer.exceptions;
-
-public class ServiceBusProcessingException extends  {
-    
-}
