--- producer/src/main/resources/jobs-config.yml
@@ -1,6 +1,0 @@
-jobs:
-  - id: thailand-routes
-    type: ROUTE_DISCOVERY
-    countryName: Thailand
-    timeToFetch: +2d
-    isActive: true
