--- producer/src/main/resources/log4j2-spring.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Configuration>
-    <Appenders>
-        <Console name="Console" target="SYSTEM_OUT">
-            <PatternLayout
-                    pattern="%style{%d{ISO8601}}{black} %highlight{%-5level }[%style{%t}{bright,blue}] %style{%C{1.}}{bright,yellow}: %msg%n%throwable" disableAnsi="false"/>
-        </Console>
-    </Appenders>
-
-    <Loggers>
-        <!-- LOG everything at INFO level -->
-        <Root level="info">
-            <AppenderRef ref="Console" />
-        </Root>
-
-        <Logger name="nomad.co.uk" level="trace" additivity="false">
-            <AppenderRef ref="Console" />
-        </Logger>
-    </Loggers>
-</Configuration>
